python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy<2.0.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2